feature/add hybrid A*

+++ include/global_planner/astar.hpp
@@ -4,9 +4,7 @@
  * @Last Modified by: Raiden49
  * @Last Modified time: 2024-06-26 10:31:22
  */
-#include <memory>
-
-#include "plan/global_planner_interface.hpp"
+#include "global_planner/global_planner_interface.hpp"
 
 namespace global_planner
 {
@@ -32,8 +30,8 @@ class AStar : public GlobalPlannerInterface {
         
         bool NeighborSearch(AstarNode& node);
         double GetNodeCost(const AstarNode& node);
-        bool GetPath(AstarNode& node, std::vector<std::array<int, 2>>& path);
-        bool GetPlan(std::vector<std::array<int, 2>>& path) override;
+        bool GetPath(AstarNode& node, std::vector<std::array<double, 2>>& path);
+        bool GetPlan(std::vector<std::array<double, 2>>& path) override;
     private:
         std::vector<AstarNode> close_set_;
         std::vector<AstarNode> open_set_;
